+++ examples/type-inference.test.ts
@@ -351,8 +351,8 @@ test("making a type from a decoder – unions", () => {
   // This “decodes” the "type" field using `getUserDecoder`. Remember that a
   // decoder is allowed to return whatever it wants. `getUserDecoder` returns a
   // _new_ decoder, which we immediately call.
-  const userDecoder5 = fields((field, obj, errors) =>
-    field("type", getUserDecoder)(obj, errors)
+  const userDecoder5 = fields((field, obj) =>
+    field("type", getUserDecoder)(obj)
   );
   type InferredType5 = ReturnType<typeof userDecoder5>;
   type ActualType5 =